requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.3
idna==2.9
multidict==4.7.6
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
